server info

+++ routes/servers.js
@@ -1,12 +1,19 @@
 const Server = require('../models/Server.js');
-
+const User = require('../models/User.js');
 
 
 module.exports = function(app) {
 
 
   app.post('/api/deleteServer', (req, res) => {
-    Server.deleteOne({_id:req.body.serverId}).exec();
+    User.findOne({_id:req.session.userId}).select({isAdmin:1}).exec((error, user) => {
+      if (user && user.isAdmin) {
+        Server.deleteOne({_id:req.body.serverId}).exec();
+        res.status(200).end();
+      } else {
+        res.status(500).send('No access.');
+      }
+    });
   })
 
   // needs auth
@@ -15,17 +22,26 @@ module.exports = function(app) {
       return res.status(500).end();
     }
 
-    Server.create({
-      name:req.body.name,
-      address:req.body.address
-    }, (error) => {
-      if (error) {
-        console.log(error);
-        res.status(500).end();
+    User.findOne({_id:req.session.userId}).select({isAdmin:1}).exec((error, user) => {
+      if (user && user.isAdmin) {
+
+        Server.create({
+          name:req.body.name,
+          address:req.body.address,
+          url: req.body.url
+        }, (error) => {
+          if (error) {
+            console.log(error);
+            res.status(500).end();
+          } else {
+            res.status(200).end();
+          }
+        })
+
       } else {
-        res.status(200).end();
+        res.status(500).send('No access.');
       }
-    })
+    });
   })
 
   app.post('/api/server', (req, res) => {
